Remove DST Root CA X3 from system certificate store
It looks like two packages provide the expired CA certificate.
The first `sed` is to remove it from `libressl`.
The second `sed` and the `rm` is to remove it from `ca-certificates`.
the upgraded `ca-certificates` is still 20161130-r3. The timestamp
still looks too old for the certificate would have been removed.